requirements.txt
sqlalchemy
openai
requests
pandas
python-dotenv
matplotlib
flask